Refaktorert bort fra deprecated Fasit consumer i AaregBehandleArbeidsforhold
