+++ apps/testnorge-aareg/aareg-core/src/main/java/no/nav/registre/aareg/consumer/ws/BehandleArbeidsforholdV1Proxy.java
@@ -1,9 +1,9 @@
 package no.nav.registre.aareg.consumer.ws;
 
-import static java.util.Objects.isNull;
-
 import lombok.RequiredArgsConstructor;
-
+import no.nav.registre.aareg.cxf.TimeoutFeature;
+import no.nav.registre.aareg.exception.TestnorgeAaregFunctionalException;
+import no.nav.registre.aareg.security.sts.StsSamlTokenService;
 import no.nav.registre.testnorge.libs.dependencyanalysis.DependencyOn;
 import no.nav.tjeneste.domene.behandlearbeidsforhold.v1.BehandleArbeidsforholdPortType;
 import org.apache.cxf.jaxws.JaxWsProxyFactoryBean;
@@ -15,9 +15,7 @@ import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
 
-import no.nav.registre.aareg.cxf.TimeoutFeature;
-import no.nav.registre.aareg.exception.TestnorgeAaregFunctionalException;
-import no.nav.registre.aareg.security.sts.StsSamlTokenService;
+import static java.util.Objects.isNull;
 
 @Component
 @RequiredArgsConstructor
@@ -31,7 +29,7 @@ public class BehandleArbeidsforholdV1Proxy {
     private static final QName BEHANDLE_ARBEIDSFORHOLD_V1 = new QName(NAMESPACE, "BehandleArbeidsforhold_v1");
 
     private final StsSamlTokenService stsSamlTokenService;
-    private final AaregBehandleArbeidsforholdFasitConsumer behandleArbeidsforholdFasitConsumer;
+    private final AaregBehandleArbeidsforhold behandleArbeidsforholdFasitConsumer;
 
     private final Map<String, BehandleArbeidsforholdPortType> wsServiceByEnvironment = new HashMap<>();
     private LocalDateTime expiry;
@@ -40,7 +38,7 @@ public class BehandleArbeidsforholdV1Proxy {
         if (hasExpired()) {
             synchronized (this) {
                 if (hasExpired()) {
-                    var urlByEnvironment = behandleArbeidsforholdFasitConsumer.fetchWsUrlsAllEnvironments();
+                    var urlByEnvironment = behandleArbeidsforholdFasitConsumer.fetchWsUrlsAllEnvironments(environment);
                     urlByEnvironment.forEach((env, url) -> wsServiceByEnvironment.put(env, createBehandleArbeidsforholdPortType(env, url)));
                     expiry = LocalDateTime.now().plusHours(4);
                 }
