+++ apps/testnorge-aareg/aareg-core/src/test/java/no/nav/registre/aareg/consumer/ws/AaregBehandleArbeidsforholdTest.java
@@ -1,0 +1,26 @@
+package no.nav.registre.aareg.consumer.ws;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+@RunWith(MockitoJUnitRunner.class)
+public class AaregBehandleArbeidsforholdTest {
+
+    @InjectMocks
+    private AaregBehandleArbeidsforhold behandleArbeidsforholdFasitConsumer;
+
+    @Test
+    public void fetchWsUrslAllEnvironments_OK() {
+        var fasitConsumers = behandleArbeidsforholdFasitConsumer.fetchWsUrlsAllEnvironments("t2");
+
+        var expectedUrl = "https://modapp-t2.adeo.no/aareg-core/BehandleArbeidsforholdService/v1";
+
+        assertThat(fasitConsumers.get("t2"), is(equalTo(expectedUrl)));
+    }
+}
